refactor: 당첨번호에서 통계 계산 제거
