+++ src/test/kotlin/lotto/WinningLottoNumberTest.kt
@@ -2,7 +2,6 @@ package lotto
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.shouldBe
-import lotto.domain.LotteryTickets
 import lotto.domain.LottoNumber
 import lotto.domain.LottoNumbers
 import lotto.domain.LottoRanking
@@ -12,50 +11,8 @@ import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.CsvSource
 import org.junit.jupiter.params.provider.MethodSource
-import java.math.BigDecimal
 
 class WinningLottoNumberTest {
-
-    @Test
-    fun `로또 번호들을 지난 주 당첨번호와 비교해서 각 등수별 개수를 알 수 있다`() {
-        val lotteryTickets = LotteryTickets(
-            listOf(
-                LottoTestHelper.makeLotteryTicket(listOf(1, 2, 3, 4, 5, 6)),
-                LottoTestHelper.makeLotteryTicket(listOf(1, 2, 3, 4, 5, 45)),
-                LottoTestHelper.makeLotteryTicket(listOf(1, 2, 3, 4, 5, 7)),
-                LottoTestHelper.makeLotteryTicket(listOf(1, 2, 3, 4, 7, 8)),
-                LottoTestHelper.makeLotteryTicket(listOf(1, 2, 3, 7, 8, 9)),
-                LottoTestHelper.makeLotteryTicket(listOf(1, 2, 7, 8, 9, 10)),
-            ),
-        )
-        val winLottoNumber = LottoTestHelper.makeWinningLottoNumbers((1..6).toList(), 45)
-
-        val result = winLottoNumber.makeRankingCountMap(lotteryTickets)
-
-        result[LottoRanking.FIRST] shouldBe 1
-        result[LottoRanking.SECOND] shouldBe 1
-        result[LottoRanking.THIRD] shouldBe 1
-        result[LottoRanking.FOURTH] shouldBe 1
-        result[LottoRanking.FIFTH] shouldBe 1
-        result[LottoRanking.MISS] shouldBe 1
-    }
-
-    @Test
-    fun `로또 번호들을 지난 주 당첨번호와 비교해서 총 수익률을 계산할 수 있다`() {
-        val lotteryTickets = LotteryTickets(
-            (1..14).map {
-                LottoTestHelper.makeLotteryTicket((it..it + 5).toList())
-            },
-        )
-        val winLottoNumber = LottoTestHelper.makeWinningLottoNumbers(
-            numbers = listOf(17, 18, 19, 20, 21, 22),
-            bonusNumber = 1,
-        )
-
-        val result = winLottoNumber.getRevenueRate(lotteryTickets)
-
-        result shouldBe BigDecimal("0.35")
-    }
 
     @ParameterizedTest(name = "matches {2}")
     @MethodSource("makeWinLottoNumber")
